Split up files.
