+++ packages/shared-ui/src/utils/top-graph-observer/index.ts
@@ -1,0 +1,7 @@
+/**
+ * @license
+ * Copyright 2024 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export { TopGraphObserver } from "./top-graph-observer.js";
