+++ packages/shared-ui/src/utils/top-graph-observer/node-entry.ts
@@ -1,0 +1,102 @@
+/**
+ * @license
+ * Copyright 2024 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import type {
+  NodeDescriptor,
+  OutputValues,
+  InputValues,
+} from "@google-labs/breadboard";
+import type {
+  RunInputEvent,
+  RunOutputEvent,
+  RunNodeStartEvent,
+} from "@google-labs/breadboard/harness";
+import { idFromPath } from "./top-graph-observer";
+import type { NodeLogEntry, ComponentActivityItem } from "../../types/types";
+
+export class NodeEntry implements NodeLogEntry {
+  type: "node";
+  id: string;
+  descriptor: NodeDescriptor;
+  hidden: boolean;
+  outputs: OutputValues | null;
+  inputs?: InputValues;
+  start: number;
+  bubbled: boolean;
+  end: number | null;
+  activity: ComponentActivityItem[] = [];
+
+  constructor(event: RunInputEvent | RunOutputEvent | RunNodeStartEvent) {
+    this.type = "node";
+    this.id = idFromPath(event.data.path);
+    this.descriptor = event.data.node;
+    this.start = event.data.timestamp;
+    this.end = null;
+
+    const type = this.descriptor.type;
+    switch (type) {
+      case "input": {
+        const inputEvent = event as RunInputEvent;
+        this.inputs = inputEvent.data.inputArguments;
+        this.bubbled = inputEvent.data.bubbled;
+        break;
+      }
+      case "output": {
+        const outputEvent = event as RunOutputEvent;
+        this.inputs = outputEvent.data.outputs;
+        this.end = event.data.timestamp;
+        this.bubbled = outputEvent.data.bubbled;
+        break;
+      }
+      default: {
+        this.bubbled = false;
+      }
+    }
+    this.outputs = null;
+    this.hidden = false;
+  }
+
+  title(): string {
+    return this.descriptor.metadata?.title || this.descriptor.id;
+  }
+}
+
+export class UserNodeEntry extends NodeEntry {
+  constructor(event: RunInputEvent) {
+    super(event);
+    this.descriptor = structuredClone(this.descriptor);
+    this.descriptor.type = "user";
+  }
+
+  title(): string {
+    return "User";
+  }
+}
+
+export class EndNodeEntry implements NodeLogEntry {
+  type = "node" as const;
+  id: string = "end";
+  activity: ComponentActivityItem[] = [];
+  descriptor = {
+    id: "end",
+    metadata: {
+      title: "End",
+    },
+    type: "end",
+  };
+  hidden = false;
+  start = globalThis.performance.now();
+  bubbled = false;
+  end = globalThis.performance.now();
+
+  constructor(reason: string) {
+    this.descriptor.metadata!.title = reason;
+  }
+
+  title(): string {
+    return this.descriptor.metadata!.title!;
+  }
+}
